LayoutTests/fast/scrolling/mac/momentum-animator-end-event-stops-expected.txt

Momentum event reached main thread
PASS scrollEventCount > 0 is true
PASS scrollEventCount is scrollEventCountAtMomentumEnded
PASS successfullyParsed is true

TEST COMPLETE
